requirements.txt
# This file was autogenerated by uv via the following command:
#    ./scripts/generate_requirements.sh
agno==2.3.12
agno-infra==1.0.4
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.3
brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
ddgs==9.6.0
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
gitdb==4.0.12
gitpython==3.1.45
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jinja2==3.1.6
jiter==0.11.0
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.2
mdurl==0.1.2
numpy==2.3.3
openai==1.109.1
packaging==25.0
pgvector==0.4.1
primp==0.15.0
psycopg==3.2.10
psycopg-binary==3.2.10
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.11.0
pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.3
pytest-cov==7.0.0
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.39.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
sqlalchemy==2.0.43
starlette==0.48.0
tomli==2.2.1
tqdm==4.67.1
typer==0.19.2
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wikipedia==1.4.0
